feat: MovieDetail 모델 타입 구현
