+++ BoxOffice/Model/Movie/DetailMovie/Actor.swift
@@ -5,4 +5,16 @@
 //  Created by 강창현 on 2/16/24.
 //
 
-import Foundation
+struct Actor: Codable {
+    let peopleName: String
+    let peopleNameEnglish: String
+    let cast: String
+    let castEnglish: String
+    
+    enum CodingKeys: String, CodingKey {
+        case peopleName = "peopleNm"
+        case peopleNameEnglish = "peopleNmEn"
+        case cast
+        case castEnglish = "castEn"
+    }
+}
