+++ BoxOffice/Model/Movie/DetailMovie/Audit.swift
@@ -5,4 +5,12 @@
 //  Created by 강창현 on 2/16/24.
 //
 
-import Foundation
+struct Audit: Codable {
+    let auditNumber: String
+    let watchGradeName: String
+    
+    enum CodingKeys: String, CodingKey {
+        case auditNumber = "auditNo"
+        case watchGradeName = "watchGradeNm"
+    }
+}
