+++ BoxOffice/Model/Movie/DetailMovie/Director.swift
@@ -5,4 +5,12 @@
 //  Created by 강창현 on 2/16/24.
 //
 
-import Foundation
+struct Director: Codable {
+    let peopleName: String
+    let peopleNameEnglish: String
+    
+    enum CodingKeys: String, CodingKey {
+        case peopleName = "peopleNm"
+        case peopleNameEnglish = "peopleNmEn"
+    }
+}
