+++ BoxOffice/Model/Movie/DetailMovie/Genre.swift
@@ -5,4 +5,10 @@
 //  Created by 강창현 on 2/16/24.
 //
 
-import Foundation
+struct Genre: Codable {
+    let genreName: String
+    
+    enum CodingKeys: String, CodingKey {
+        case genreName = "genreNm"
+    }
+}
