+++ BoxOffice/Model/Movie/DetailMovie/MovieInfoResult.swift
@@ -5,4 +5,7 @@
 //  Created by 강창현 on 2/16/24.
 //
 
-import Foundation
+struct MovieInfoResult: Codable {
+    let movieInfo: MovieInfo
+    let source: String
+}
