+++ BoxOffice/Model/Movie/DetailMovie/Nation.swift
@@ -5,4 +5,10 @@
 //  Created by 강창현 on 2/16/24.
 //
 
-import Foundation
+struct Nation: Codable {
+    let nationName: String
+    
+    enum CodingKeys: String, CodingKey {
+        case nationName = "nationNm"
+    }
+}
