+++ BoxOffice/Model/Movie/DetailMovie/ShowType.swift
@@ -5,4 +5,12 @@
 //  Created by 강창현 on 2/16/24.
 //
 
-import Foundation
+struct ShowType: Codable {
+    let showTypeGroupName: String
+    let showTypeName: String
+    
+    enum CodingKeys: String, CodingKey {
+        case showTypeGroupName = "showTypeGroupNm"
+        case showTypeName = "showTypeNm"
+    }
+}
